% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_status}
\alias{check_status}
\title{Helper function to check the status of a response object}
\usage{
check_status(res)
}
\arguments{
\item{res}{A httr response object.}
}
\description{
Parses the error number and message
}
