% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_break.R
\name{find_break}
\alias{find_break}
\title{Calculate breakpoints}
\usage{
find_break(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the index of breakpoint or NA
}
\description{
Calculate breakpoints
}
