% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_and_parse}
\alias{get_and_parse}
\title{Helper function to query an URL}
\usage{
get_and_parse(url)
}
\arguments{
\item{url}{Character vector of the url}
}
\value{
The parsed content of the response. An error message in cases the
GET request fails
}
\description{
Helper function to query an URL
}
\keyword{internal}
