% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{wapor_collections}
\alias{wapor_collections}
\title{List collections at FAO GISManager}
\usage{
wapor_collections()
}
\value{
A data.frame object with code and captions of the available collections.
}
\description{
This function lists available data collections from the FAO GISManager API.
Only the code name used for other functions and the caption (e.g.) long name
is returned. This function is not equivalent to \code{wapor_collection()}
which returns the metadata of a \emph{singular} data collection.
}
