% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_indices.R
\name{wapor_extract}
\alias{wapor_extract}
\title{Extract and summarize index values per polygon}
\usage{
wapor_extract(input_files, aoi)
}
\arguments{
\item{input_files}{A charachter vector pointing to a complete set of
output files from the wapor_indices() function}

\item{aoi}{A sf-object with optionally multiple polygons for which to extract
the index information.}
}
\value{
A list of tibble objects. See the decription for details.
}
\description{
This function can be used to extract and summarise the data output by
the wapor_indices function on a polygon basis. The data is returned
as a list object with several tables in long-format. These objects are:
\itemize{
\item Land Cover Statistics (lcc_stats): breakpoint, median land use, modal land use
\item Land Cover Frequency (lcc_freq): the averaged frequency of occurence for selected land cover classes
\item Indices (indices): Averaged yearly statistics of the raw parameters as well as LP1_global and LP1_yearly
\item Double Season: Class specific occurence of double seasons per year
\item Double Season Difference: Difference in number of double season in pre- and post epoch
\item Epoch Differences: Differences in productivity for raw parameters
\item Trends: Estimates and p-values of a linear trend per parameter
}
}
