% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{wapor_metadata}
\alias{wapor_metadata}
\title{FAO GIS get product metadata}
\usage{
wapor_metadata(collection, product)
}
\arguments{
\item{collection}{A length one character vector specifying the collection.}

\item{product}{A length one character vector specifying the product..}
}
\value{
A list object with the obhects \code{info}, \code{dimensions}, and \code{meta}.
}
\description{
This function retrieves metadata of a specific product within a collection.
The returned object is a three-leveled list with the first object containing
some general information on the product. The second object contains information
of the dimensions the data set is associated with wile the third element contains
additional metadata e.g. on spatio-temporal resolution or the methodology
how the product was generated.
}
