% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nbwp.R
\name{wapor_nbwp}
\alias{wapor_nbwp}
\title{Calculate Net Biomass Water Productivity}
\usage{
wapor_nbwp(
  input_files = NULL,
  years = NULL,
  aoi = NULL,
  verbose = TRUE,
  outdir = "."
)
}
\arguments{
\item{input_files}{A character vector with the file paths to the seasonal
phenological metrics (PHE), the decadal transpiration (T), the seasonal
total biomass production (TBP) and possibly the annual land cover classification (LCC).}

\item{years}{The years for which to calculate NBWP as a numeric vector. Note,
NBWP can be calculated only for years for which data of a previous and a
subsequent years are available. Usually that means that the earliest possible
starting year is 2010 because data availability starts in 2009. The latest
possible year for that NBWP can be calculated is usually two years
before the current year.}

\item{aoi}{An optional sf object that can be used to set the calculation of
NBWP to its bounding box.}

\item{verbose}{A logical indicating if informative messages should be printed.}

\item{outdir}{A character vector indicating the directory where the output shall
be written to. The function will create two new sub-directories (T_S and NBWP)
and overwrite potentially already available files in these sub-directories.}
}
\value{
A character vector with the file locations of the NBWP rasters.
}
\description{
This function is used to calculate the seasonal net water productivity
based on the phenologocial information (PHE) and the seasonal total
biomass production (TBP). Additionally, a LCC land use mask can be used
to restrict the analysis to pixels of a certain class. The function writes
the seasonal total sum of transpiration (T_S) and the net biomass productivity
(NBWP) to respective directories in the output directory.
}
