% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{wapor_products}
\alias{wapor_products}
\title{FAo GIS List available Products in collection}
\usage{
wapor_products(collection)
}
\arguments{
\item{collection}{A character vector of length one specifying the collection
for which to retrieve available products-}
}
\value{
A list object with all available products. One product object contains
of the two data.frames \code{product} and \code{meta} where meta can have
variable column number due to differences in the available meta data.
}
\description{
This function lists all available datasets within a collection. The result
is returned as a list where each element represents a distinct product and is
named accordingly. One level lower, there are two objects one containing general
information on the product and the other one containing additional metadata,
such as spatio-temporal resolution, extent in space and time, coordinate system etc.
}
