% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wapor_signin}
\alias{wapor_signin}
\title{Sign in with API key at FAO GISManager to retrieve access token}
\usage{
wapor_signin(APIkey)
}
\arguments{
\item{APIkey}{A length one character containing a user-specific API key.}
}
\value{
An access token.
}
\description{
Function is used internally to sign in and parse the access_token.
}
\keyword{internal}
